Move workgroup passes & tests
