+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/BUILD.bazel
@@ -34,19 +34,17 @@ iree_compiler_cc_library(
         "Canonicalizer.cpp",
         "CaptureDynamicDims.cpp",
         "CleanupTensorShapes.cpp",
-        "ConvertDispatchRegionsToWorkgroups.cpp",
         "ConvertMeshToFlow.cpp",
         "ConvertRegionToWorkgroups.cpp",
-        "ConvertTensorToFlow.cpp",
         "ConvertToFlow.cpp",
         "DeduplicateExecutables.cpp",
         "DumpDispatchGraph.cpp",
         "ExportBenchmarkFuncs.cpp",
+        "FormDispatchRegions.cpp",
         "InitializeEmptyTensors.cpp",
         "InjectDispatchTracing.cpp",
         "InjectTensorTracing.cpp",
         "InsertDispatchDebugTargets.cpp",
-        "MaterializeDefaultWorkgroupCountRegion.cpp",
         "OutlineConstants.cpp",
         "OutlineDispatchExterns.cpp",
         "OutlineDispatchRegions.cpp",
@@ -57,6 +55,7 @@ iree_compiler_cc_library(
     ],
     hdrs = [
         "ConvertRegionToWorkgroups.h",
+        "FormDispatchRegions.h",
         "Passes.h",
         "Passes.h.inc",
         "RegionOpUtils.h",
