+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/CMakeLists.txt
@@ -24,6 +24,7 @@ iree_cc_library(
     Transforms
   HDRS
     "ConvertRegionToWorkgroups.h"
+    "FormDispatchRegions.h"
     "Passes.h"
     "Passes.h.inc"
     "RegionOpUtils.h"
@@ -32,19 +33,17 @@ iree_cc_library(
     "Canonicalizer.cpp"
     "CaptureDynamicDims.cpp"
     "CleanupTensorShapes.cpp"
-    "ConvertDispatchRegionsToWorkgroups.cpp"
     "ConvertMeshToFlow.cpp"
     "ConvertRegionToWorkgroups.cpp"
-    "ConvertTensorToFlow.cpp"
     "ConvertToFlow.cpp"
     "DeduplicateExecutables.cpp"
     "DumpDispatchGraph.cpp"
     "ExportBenchmarkFuncs.cpp"
+    "FormDispatchRegions.cpp"
     "InitializeEmptyTensors.cpp"
     "InjectDispatchTracing.cpp"
     "InjectTensorTracing.cpp"
     "InsertDispatchDebugTargets.cpp"
-    "MaterializeDefaultWorkgroupCountRegion.cpp"
     "OutlineConstants.cpp"
     "OutlineDispatchExterns.cpp"
     "OutlineDispatchRegions.cpp"
