+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/test/BUILD.bazel
@@ -20,15 +20,9 @@ iree_lit_test_suite(
             "capture_dispatch_dynamic_dims.mlir",
             "capture_scf_for_dynamic_dims.mlir",
             "cleanup_tensor_shapes.mlir",
-            "convert_region_to_workgroups.mlir",
             "deduplicate_executables.mlir",
             "export_benchmark_funcs.mlir",
-            "dispatch_linalg_ext_fusion.mlir",
-            "dispatch_linalg_on_tensors.mlir",
-            "dispatch_linalg_on_tensors_default.mlir",
-            "dispatch_linalg_on_tensors_fusion_with_transpose.mlir",
             "flow_canonicalize.mlir",
-            "form_dispatch_workgroups.mlir",
             "initialize_empty_tensors.mlir",
             "inject_dispatch_tracing.mlir",
             "inject_tensor_tracing.mlir",
