+++ compiler/src/iree/compiler/DispatchCreation/BUILD.bazel
@@ -19,6 +19,8 @@ iree_compiler_cc_library(
         "CloneProducersIntoDispatchRegions.cpp",
         "CollapseDimensions.cpp",
         "CollapseReductionDimensions.cpp",
+        "ConvertDispatchRegionsToWorkgroups.cpp",
+        "ConvertTensorToFlow.cpp",
         "DispatchWithTransformDialect.cpp",
         "ElementwiseOpFusion.cpp",
         "FoldUnitExtentDims.cpp",
@@ -28,6 +30,7 @@ iree_compiler_cc_library(
         "FuseMultiUseElementwiseProducer.cpp",
         "FusionPreprocessing.cpp",
         "FusionUtils.cpp",
+        "MaterializeDefaultWorkgroupCountRegion.cpp",
         "Passes.cpp",
         "SetEncoding.cpp",
         "SinkReshapes.cpp",
@@ -36,7 +39,6 @@ iree_compiler_cc_library(
         "TransposeGenericOps.cpp",
     ],
     hdrs = [
-        "FormDispatchRegions.h",
         "FusionUtils.h",
         "Passes.h",
     ],
