+++ compiler/src/iree/compiler/DispatchCreation/FormDispatchRegions.cpp
@@ -4,8 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/compiler/DispatchCreation/FormDispatchRegions.h"
-
+#include "iree/compiler/Dialect/Flow/Transforms/FormDispatchRegions.h"
 #include "iree/compiler/Dialect/Encoding/IR/EncodingOps.h"
 #include "iree/compiler/Dialect/Flow/IR/FlowDialect.h"
 #include "iree/compiler/Dialect/Flow/IR/FlowOps.h"
@@ -46,79 +45,10 @@
 static const char kRootOpAttr[] = "__root_op__";
 static const char kFusionGroupsAttr[] = "__fused_op__";
 
-//===----------------------------------------------------------------------===//
-// Definition of TensorDimTrackingRewriter
-//===----------------------------------------------------------------------===//
-
-namespace mlir {
-
-TensorDimTrackingRewriter::TensorDimTrackingRewriter(Operation *op)
-    : IRRewriter(op->getContext()) {
-  setListener(this);
-  op->walk([&](tensor::DimOp dimOp) { dimOps.insert(dimOp.getOperation()); });
-}
-SmallVector<tensor::DimOp> TensorDimTrackingRewriter::getTensorDimOps() {
-  SmallVector<tensor::DimOp> result;
-  for (Operation *op : dimOps)
-    result.push_back(cast<tensor::DimOp>(op));
-  return result;
-}
-void TensorDimTrackingRewriter::notifyOperationErased(Operation *op) {
-  IRRewriter::Listener::notifyOperationErased(op);
-  if (isa<tensor::DimOp>(op))
-    dimOps.erase(op);
-}
-
-void TensorDimTrackingRewriter::notifyOperationInserted(Operation *op,
-                                                        InsertPoint previous) {
-  IRRewriter::Listener::notifyOperationInserted(op, previous);
-  if (isa<tensor::DimOp>(op))
-    dimOps.insert(op);
-}
-
-} // namespace mlir
-
 namespace mlir::iree_compiler::DispatchCreation {
 
 #define GEN_PASS_DEF_FORMDISPATCHREGIONSPASS
 #include "iree/compiler/DispatchCreation/Passes.h.inc"
-
-LogicalResult simplifyDimOps(RewriterBase &rewriter,
-                             const SmallVector<tensor::DimOp> &dimOps) {
-  for (tensor::DimOp dimOp : dimOps) {
-    // Only DimOps with static indices are supported.
-    std::optional<int64_t> idx = dimOp.getConstantIndex();
-    if (!idx.has_value())
-      continue;
-    // Only DimOps with ranked tensors are supported.
-    auto tensorType =
-        llvm::dyn_cast<RankedTensorType>(dimOp.getSource().getType());
-    if (!tensorType)
-      continue;
-
-    if (!tensorType.isDynamicDim(*idx)) {
-      // Rewrite static dimension with constant.
-      OpBuilder::InsertionGuard g(rewriter);
-      rewriter.setInsertionPoint(dimOp);
-      int64_t size = tensorType.getShape()[*idx];
-      rewriter.replaceOpWithNewOp<arith::ConstantIndexOp>(dimOp, size);
-      continue;
-    }
-
-    // Try to simplify dynamic dims.
-    SmallVector<Value> dynamicDims;
-    if (succeeded(IREE::Flow::getOptimizedDynamicResultDims(
-            rewriter, dimOp.getSource(), dynamicDims))) {
-      unsigned ctr = 0;
-      for (int64_t i = 0; i < *dimOp.getConstantIndex(); ++i)
-        if (tensorType.isDynamicDim(i))
-          ++ctr;
-      rewriter.replaceOp(dimOp, dynamicDims[ctr]);
-    }
-  }
-
-  return success();
-}
 
 //===----------------------------------------------------------------------===//
 // Root and fusion group attribute handling
@@ -941,7 +871,7 @@ createFusionGroups(TensorDimTrackingRewriter &rewriter,
     // Simplify tensor::DimOps.
     {
       SmallVector<tensor::DimOp> dimOps = rewriter.getTensorDimOps();
-      if (failed(simplifyDimOps(rewriter, dimOps))) {
+      if (failed(IREE::Flow::simplifyDimOps(rewriter, dimOps))) {
         return failure();
       }
     }
@@ -965,7 +895,7 @@ createFusionGroups(TensorDimTrackingRewriter &rewriter,
       // Simplify tensor::DimOps.
       {
         SmallVector<tensor::DimOp> dimOps = rewriter.getTensorDimOps();
-        if (failed(simplifyDimOps(rewriter, dimOps))) {
+        if (failed(IREE::Flow::simplifyDimOps(rewriter, dimOps))) {
           return failure();
         }
       }
@@ -979,7 +909,7 @@ createFusionGroups(TensorDimTrackingRewriter &rewriter,
     // Simplify tensor::DimOps.
     {
       SmallVector<tensor::DimOp> dimOps = rewriter.getTensorDimOps();
-      if (failed(simplifyDimOps(rewriter, dimOps))) {
+      if (failed(IREE::Flow::simplifyDimOps(rewriter, dimOps))) {
         return failure();
       }
     }
