+++ compiler/src/iree/compiler/DispatchCreation/test/BUILD.bazel
@@ -25,6 +25,12 @@ iree_lit_test_suite(
             "dispatch_region_formation_preprocessing.mlir",
             "fold_unit_dims.mlir",
             "form_dispatch_regions.mlir",
+            "dispatch_linalg_on_tensors.mlir",
+            "convert_region_to_workgroups.mlir",
+            "form_dispatch_workgroups.mlir",
+            "dispatch_linalg_ext_fusion.mlir",
+            "dispatch_linalg_on_tensors_default.mlir",
+            "dispatch_linalg_on_tensors_fusion_with_transpose.mlir",
             "form_scalar_dispatches.mlir",
             "fuse_horizontal_contractions.mlir",
             "fuse_multiuse_elementwise_producer.mlir",
