+++ compiler/src/iree/compiler/DispatchCreation/test/CMakeLists.txt
@@ -19,10 +19,16 @@ iree_lit_test_suite(
     "collapse_dimensions.mlir"
     "collapse_linalg_generic_on_tensors.mlir"
     "collapse_reduction.mlir"
+    "convert_region_to_workgroups.mlir"
+    "dispatch_linalg_ext_fusion.mlir"
+    "dispatch_linalg_on_tensors.mlir"
+    "dispatch_linalg_on_tensors_default.mlir"
+    "dispatch_linalg_on_tensors_fusion_with_transpose.mlir"
     "dispatch_linalg_transform_dialect.mlir"
     "dispatch_region_formation_preprocessing.mlir"
     "fold_unit_dims.mlir"
     "form_dispatch_regions.mlir"
+    "form_dispatch_workgroups.mlir"
     "form_scalar_dispatches.mlir"
     "fuse_horizontal_contractions.mlir"
     "fuse_multiuse_elementwise_producer.mlir"
